Split out Objects.hs from QC.hs
This is the first file split out from QC.hs - an easy one, since it
has just one test.

The patch changes the way we build hpc excludes, since now we'll have
many modules that need to be excluded, and hpc doesn't seem to be able
to do wildcards (it can do packages, but all our code is
not-package-based).

Further splitting will be done in bigger batches.

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: René Nussbaumer <rn@google.com>